+++ catalyst-gateway/tests/api_tests/api_tests/__init__.py
@@ -4,10 +4,7 @@ import asyncio
 import asyncpg
 import http.client
 
-GET_VERSION_QUERY = "SELECT MAX(version) FROM refinery_schema_history"
-UPDATE_QUERY = "UPDATE refinery_schema_history SET version=$1 WHERE version=$2"
-
-DB_URL="postgres://catalyst-event-dev:CHANGE_ME@event-db/CatalystEventDev"
+DB_URL = "postgres://catalyst-event-dev:CHANGE_ME@event-db/CatalystEventDev"
 DEFAULT_TIMEOUT: int = 10
 HOST = "gateway"
 PORT = 3030
@@ -19,27 +16,3 @@ def call_api_url(method, endpoint, *args, **kwargs):
     client.close()
     return resp
 
-def fetch_schema_version():
-    async def get_current_version():
-        conn = await asyncpg.connect(DB_URL)
-        if conn is None:
-            raise Exception("no db connection found")
-
-        version = await conn.fetchval(GET_VERSION_QUERY)
-        if version is None:
-            raise Exception("failed to fetch version from db")
-        return version
-
-    return asyncio.run(get_current_version())
-
-def change_version(from_value: int, change_to: int):
-    async def change_schema_version():
-        conn = await asyncpg.connect(DB_URL)
-        if conn is None:
-            raise Exception("no db connection found for")
-
-        update = await conn.execute(UPDATE_QUERY, change_to, from_value)
-        if update is None:
-            raise Exception("failed to fetch version from db")
-
-    return asyncio.run(change_schema_version())
