+++ catalyst-gateway/tests/api_tests/api_tests/test_schema_mismatch_behavior.py
@@ -1,8 +1,39 @@
 """Test the `catalyst-gateway` service when a DB schema mismatch occurs."""
 
 from loguru import logger
+import requests
 
-from api_tests import call_api_url, fetch_schema_version, change_version
+from api_tests import DB_URL, call_api_url
+
+GET_VERSION_QUERY = "SELECT MAX(version) FROM refinery_schema_history"
+UPDATE_QUERY = "UPDATE refinery_schema_history SET version=$1 WHERE version=$2"
+
+
+def fetch_schema_version():
+    async def get_current_version():
+        conn = await asyncpg.connect(DB_URL)
+        if conn is None:
+            raise Exception("no db connection found")
+
+        version = await conn.fetchval(GET_VERSION_QUERY)
+        if version is None:
+            raise Exception("failed to fetch version from db")
+        return version
+
+    return asyncio.run(get_current_version())
+
+
+def change_version(from_value: int, change_to: int):
+    async def change_schema_version():
+        conn = await asyncpg.connect(DB_URL)
+        if conn is None:
+            raise Exception("no db connection found for")
+
+        update = await conn.execute(UPDATE_QUERY, change_to, from_value)
+        if update is None:
+            raise Exception("failed to fetch version from db")
+
+    return asyncio.run(change_schema_version())
 
 
 def check_is_live():
